A unidirectional experiment
This is an attempt at a vanilla unidirectional framework. It doesn't do much
yet, but it does have components that render based on their props and actions,
and those actions get propagated up at least one level. So, kinda?

For the sake of expediency, I am going back to vanilla MVC for the time being.