HTML skeleton for an editor
This looks like a maze editor, but it doesn't really do anything. This proves
that a maze can be rendered in an editor using flexbox and divs. I chose divs
over buttons, to avoid weird focus styles. However, we do lose keyboard
functionality with that. Right now, this only handles clicks. Drag-drop
behavior would be nice.