Add ability to toggle tiles
You can toggle tiles now, which dispatches an event through the unidirectional
architecture. There are some flaws at the moment. For one, all of the tiles
rerender after every click. This could be a performance issue eventually,
although it seems OK for now, on a 9x9 maze. Another flaw is that the store
is being passed into the MazeComponent, and the MazeComponent and TileComponent
are not the same structure. I believe these go hand in hand (mounting requires
some state, which the Tiles will likely have eventually to avoid rerendering).