Fixed collision detection, I think
This improves collision detection to make sure that you don't step into a
wall. It checks where you are actually going to step and has cases for each
kind of tile next to you and in front of you. There is definitely some
cleanup that would make this easier to follow.