Convert remaining behavior to custom element
There is definitely some hackiness in here, but it is all working as
part of a custom element! There are issues with lighting (and maybe
ground?) when you render two at once.

There is also work to be done to make the win condition an event rather
than an alert, and probably to try to protect some data a little more.