Build a simple REST backend for mazes
This adds a simple REST server to save mazes to MongoDB.
There are some pieces that I'd like to clean up around
deprecation warnings and the like, but for now this works.

This also removes a bunch of code from the THREE renderer,
which has since been copied to the component directory.