Move files to new clean app directory
These files are all broken right now, because they don't have an
API to load data with. This signals the end of Auth as a use case.
Now these files can serve as placeholders for the beginning of
other use case driven development, like:

* Creating a new user and their mazes
* Loading your list of mazes when you log in
* Editing a maze draft
* Running a maze draft

This also contains a small change to the email token URL, so that
works again.