Add a mock data file
This commit adds three users and six mazes (with a handful of revisions).
This is the stepping stone for mockable repositories, which is itself a
stepping stone to better unit tests for use cases.

With the ability to create testable use cases, I can focus on only the
queries and data modifications I need. Then, I can plug those use cases
directly into the front end. This way, I only need to deal with the web
folder when I want to render a page (or trigger a use case based on user
action.