Implement PopulateUser usecase
This implements (reimplements?) the use case to populate user info
after authenticating. This will happen on most requests. To follow
functional patterns, it takes in a verified email and userRepo and
resolves to the populated user information *and* a userRepo. This
covers the case where a user is newly created, and is mostly done
so that our test userRepo does not have to clean up state between
unit tests.

The mock user_repository methods needed for this usecase do not
exactly follow the behavior that the mongodb user repository does.
Possibly start there next time (or pick a new use case to fix up).