Implement userRepo.readByEmail and .add for mongo
This implements the readByEmail and add methods for the user repo
that is backed by MongoDB. This is mostly existing behavior, but
built to match a testable interface. Note, however, that this is
not actually an immutable structure, even though mutations like
add will return a "new" reference to the same repository.

The populateUser usecase is a little different than the previous
addUser scenario, and so this commit also disables some tests and
comments out some code that handles that. This likely (certainly)
breaks the app... but for now we are focusing on individual cases.