Add ability to run mazes in challenges!
Challenges can be updated with your run time.

Caveats:  The views are very fragile and give you links to do things
that you cannot do given the challenge state. Also, many actions are
not properly guarded against, in case a bad actor tried to hijack a
challenge. These steps are definitely next, but this is exciting!

This is perhaps the most critical piece of functionality, and
it is done!!!